+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v8.43.3 (unreleased)
 ### Implementation changes and bug fixes
+- [PR #2206](https://github.com/rqlite/rqlite/pull/2206): Add convenience WriteOne() to Queue.
 - [PR #2204](https://github.com/rqlite/rqlite/pull/2204): Add byte-marshalling for CDC Events.
 - [PR #2205](https://github.com/rqlite/rqlite/pull/2205): Improve CDC Protobuf naming.